Slight Plant Tweaks
I added some grass to the bases of the Roofed Oak and Acacia saplings.

I wanted some opinions on the tiny changes I made to the flowers.  I
added a bit more texture to the peonies and edited the allium bloom so
it looked a bit more like the huge ball of flowers that they are.